:recycle: 채팅방 조회 비즈니스 리팩토링
