+++ src/main/kotlin/com/stark/shoot/application/service/chatroom/FindChatroomService.kt
@@ -1,11 +1,10 @@
 package com.stark.shoot.application.service.chatroom
 
 import com.stark.shoot.adapter.`in`.web.dto.chatroom.ChatRoomResponse
+import com.stark.shoot.adapter.`in`.web.mapper.ChatRoomResponseMapper
 import com.stark.shoot.application.port.`in`.chatroom.FindChatRoomUseCase
 import com.stark.shoot.application.port.out.chatroom.LoadChatRoomPort
-import com.stark.shoot.domain.chat.room.ChatRoom
 import com.stark.shoot.domain.chat.room.service.ChatRoomDomainService
-import com.stark.shoot.adapter.`in`.web.mapper.ChatRoomResponseMapper
 import com.stark.shoot.infrastructure.annotation.UseCase
 import org.springframework.transaction.annotation.Transactional
 
@@ -30,9 +29,9 @@ class FindChatroomService(
         val chatRooms = loadChatRoomPort.findByParticipantId(userId)
 
         // 채팅방 정보 준비
-        val titles = prepareChatRoomTitles(chatRooms, userId)
-        val lastMessages = prepareLastMessages(chatRooms)
-        val timestamps = prepareTimestamps(chatRooms)
+        val titles = chatRoomDomainService.prepareChatRoomTitles(chatRooms, userId)
+        val lastMessages = chatRoomDomainService.prepareLastMessages(chatRooms)
+        val timestamps = chatRoomDomainService.prepareTimestamps(chatRooms)
 
         // 채팅방 목록을 ChatRoomResponse로 변환하여 반환합니다.
         return chatRoomResponseMapper.toResponseList(chatRooms, userId, titles, lastMessages, timestamps)
@@ -53,48 +52,16 @@ class FindChatroomService(
         val chatRooms = loadChatRoomPort.findByParticipantId(userId1)
 
         // 두 사용자 간의 1:1 채팅방을 찾습니다. (도메인 객체의 정적 메서드 사용)
-        val directChatRoom = ChatRoom.findDirectChatBetween(chatRooms, userId1, userId2)
+        val directChatRoom = chatRoomDomainService.findDirectChatBetween(chatRooms, userId1, userId2)
 
         // 채팅방이 없으면 null 반환
-        return directChatRoom?.let { 
+        return directChatRoom?.let {
             val title = it.createChatRoomTitle(userId1)
             val lastMessage = it.createLastMessageText()
             val timestamp = it.formatTimestamp()
 
-            chatRoomResponseMapper.toResponse(it, userId1, title, lastMessage, timestamp) 
+            chatRoomResponseMapper.toResponse(it, userId1, title, lastMessage, timestamp)
         }
     }
 
-    /**
-     * 채팅방 제목 맵 준비
-     */
-    private fun prepareChatRoomTitles(chatRooms: List<ChatRoom>, userId: Long): Map<Long, String> {
-        return chatRooms.associate { room ->
-            val roomId = room.id ?: 0L
-            val title = room.createChatRoomTitle(userId)
-            roomId to title
-        }
-    }
-
-    /**
-     * 마지막 메시지 맵 준비
-     */
-    private fun prepareLastMessages(chatRooms: List<ChatRoom>): Map<Long, String> {
-        return chatRooms.associate { room ->
-            val roomId = room.id ?: 0L
-            val lastMessage = room.createLastMessageText()
-            roomId to lastMessage
-        }
-    }
-
-    /**
-     * 타임스탬프 맵 준비
-     */
-    private fun prepareTimestamps(chatRooms: List<ChatRoom>): Map<Long, String> {
-        return chatRooms.associate { room ->
-            val roomId = room.id ?: 0L
-            val timestamp = room.formatTimestamp()
-            roomId to timestamp
-        }
-    }
 }
